+++ rats-apps/src/rats/annotations/__init__.py
@@ -4,22 +4,6 @@ General purpose library to attach annotations to functions.
 Annotations are typically, but not exclusively, attached using decorators.
 """
 
-from ._functions import (
-    AnnotationsContainer,
-    DecoratorType,
-    GroupAnnotations,
-    T_GroupType,
-    annotation,
-    get_annotations,
-    get_class_annotations,
-)
+import lazy_loader
 
-__all__ = [
-    "AnnotationsContainer",
-    "DecoratorType",
-    "GroupAnnotations",
-    "T_GroupType",
-    "annotation",
-    "get_annotations",
-    "get_class_annotations",
-]
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)  # type: ignore[reportUnknownVariableType]
