+++ rats-apps/src/rats/annotations/__init__.pyi
@@ -1,0 +1,19 @@
+from ._functions import (
+    AnnotationsContainer,
+    DecoratorType,
+    GroupAnnotations,
+    T_GroupType,
+    annotation,
+    get_annotations,
+    get_class_annotations,
+)
+
+__all__ = [
+    "AnnotationsContainer",
+    "DecoratorType",
+    "GroupAnnotations",
+    "T_GroupType",
+    "annotation",
+    "get_annotations",
+    "get_class_annotations",
+]
