+++ rats-apps/src/rats/logs/__init__.pyi
@@ -1,0 +1,5 @@
+from ._app import ConfigureApplication
+
+__all__ = [
+    "ConfigureApplication",
+]
