+++ rats-apps/src/rats_e2e/apps/__init__.py
@@ -1,7 +1,5 @@
 """Examples and end-to-end tests for the `rats.apps` package."""
 
-from ._example_cli import ExampleCliApp
+import lazy_loader
 
-__all__ = [
-    "ExampleCliApp",
-]
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)  # type: ignore[reportUnknownVariableType]
