+++ rats-apps/src/rats_e2e/apps/__init__.pyi
@@ -1,0 +1,5 @@
+from ._example_cli import ExampleCliApp
+
+__all__ = [
+    "ExampleCliApp",
+]
