+++ rats-apps/src/rats_e2e/apps/inputs/__init__.pyi
@@ -1,0 +1,8 @@
+from ._app import AppInput, Application, AppServices, main
+
+__all__ = [
+    "AppInput",
+    "AppServices",
+    "Application",
+    "main",
+]
