+++ rats-apps/src/rats_e2e/apps/minimal/__init__.py
@@ -18,9 +18,6 @@ or directly through a terminal.
     ```
 """
 
-from ._app import Application, main
+import lazy_loader
 
-__all__ = [
-    "Application",
-    "main",
-]
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)  # type: ignore[reportUnknownVariableType]
