+++ rats-apps/src/rats_e2e/apps/minimal/__init__.pyi
@@ -1,0 +1,6 @@
+from ._app import Application, main
+
+__all__ = [
+    "Application",
+    "main",
+]
