+++ rats-apps/test/rats_test/apps/example/__init__.py
@@ -1,28 +1,3 @@
-from ._app import ExampleApp
-from ._dummy_containers import DummyContainerServiceIds
-from ._example_fallbacks import (
-    ExampleFallbackPlugin1,
-    ExampleFallbackPlugin2,
-    ExampleFallbackPlugin3,
-)
-from ._example_groups import ExampleGroupsPlugin1, ExampleGroupsPlugin2
-from ._ids import ExampleConfigIds, ExampleIds
-from ._storage import StorageClient, StorageSettings
-from ._storage_plugin import ExampleStoragePlugin
+import lazy_loader
 
-__all__ = [
-    "DummyContainerServiceIds",
-    "ExampleApp",
-    "ExampleConfigIds",
-    "ExampleConfigIds",
-    "ExampleFallbackPlugin1",
-    "ExampleFallbackPlugin2",
-    "ExampleFallbackPlugin3",
-    "ExampleGroupsPlugin1",
-    "ExampleGroupsPlugin2",
-    "ExampleIds",
-    "ExampleIds",
-    "ExampleStoragePlugin",
-    "StorageClient",
-    "StorageSettings",
-]
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)  # type: ignore[reportUnknownVariableType]
