+++ rats-devtools/src/rats/aml/__init__.py
@@ -1,24 +1,5 @@
 """Submit applications as aml jobs, sending necessary context from the driver node."""
 
-from ._app import AppConfigs, Application, AppServices, main
-from ._configs import (
-    AmlEnvironment,
-    AmlIO,
-    AmlJobContext,
-    AmlJobDetails,
-    AmlWorkspace,
-)
-from ._submission import submit
+import lazy_loader
 
-__all__ = [
-    "AmlEnvironment",
-    "AmlIO",
-    "AmlJobContext",
-    "AmlJobDetails",
-    "AmlWorkspace",
-    "AppConfigs",
-    "AppServices",
-    "Application",
-    "main",
-    "submit",
-]
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
