+++ rats-devtools/src/rats/docs/__init__.pyi
@@ -1,0 +1,8 @@
+from ._app import AppConfigs, Application, AppServices, main
+
+__all__ = [
+    "AppConfigs",
+    "AppServices",
+    "Application",
+    "main",
+]
