+++ rats-devtools/src/rats/ez/__init__.pyi
@@ -1,0 +1,6 @@
+from ._app import Application, main
+
+__all__ = [
+    "Application",
+    "main",
+]
