+++ rats-devtools/src/rats/projects/__init__.py
@@ -5,24 +5,6 @@ A project is loosely tied to one repository, which is made up of one or more com
 component is a separate entity that can be built, tested, and released independently.
 """
 
-from ._component_tools import ComponentId, ComponentTools, UnsetComponentTools
-from ._plugin import PluginContainer, PluginServices, find_repo_root
-from ._project_tools import (
-    ComponentNotFoundError,
-    ProjectConfig,
-    ProjectNotFoundError,
-    ProjectTools,
-)
+import lazy_loader
 
-__all__ = [
-    "ComponentId",
-    "ComponentNotFoundError",
-    "ComponentTools",
-    "PluginContainer",
-    "PluginServices",
-    "ProjectConfig",
-    "ProjectNotFoundError",
-    "ProjectTools",
-    "UnsetComponentTools",
-    "find_repo_root",
-]
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
