+++ rats-devtools/src/rats/projects/__init__.pyi
@@ -1,0 +1,21 @@
+from ._component_tools import ComponentId, ComponentTools, UnsetComponentTools
+from ._plugin import PluginContainer, PluginServices, find_repo_root
+from ._project_tools import (
+    ComponentNotFoundError,
+    ProjectConfig,
+    ProjectNotFoundError,
+    ProjectTools,
+)
+
+__all__ = [
+    "ComponentId",
+    "ComponentNotFoundError",
+    "ComponentTools",
+    "PluginContainer",
+    "PluginServices",
+    "ProjectConfig",
+    "ProjectNotFoundError",
+    "ProjectTools",
+    "UnsetComponentTools",
+    "find_repo_root",
+]
