+++ rats-devtools/src/rats_e2e/aml/__init__.pyi
@@ -1,0 +1,8 @@
+from ._app import Application, main
+from ._example_job import ExampleJob
+
+__all__ = [
+    "Application",
+    "ExampleJob",
+    "main",
+]
