+++ rats-devtools/test/rats_test_resources/projects/example-pdm/src/example_pdm/__init__.py
@@ -1,0 +1,3 @@
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
