+++ rats-devtools/test/rats_test_resources/projects/example-uv/src/example_uv/__init__.py
@@ -1,2 +1,3 @@
-def main() -> None:
-    print("Hello from example-uv!")
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
