+++ rats-devtools/test/rats_test_resources/projects/example-uv/src/example_uv/__init__.pyi
@@ -1,0 +1,2 @@
+def main() -> None:
+    print("Hello from example-uv!")
